shorten minfreeHeap json field name in info object

+++ include/application.h
@@ -88,7 +88,7 @@ public:
     uint32_t getUptime();
     void uptimeCounter();
     size_t getFreeHeapSize();
-    bool checkHeap(size_t minHeap);
+    bool checkHeap(uint32_t minHeap);
     size_t getMinimumHeapUptime() { return _minimumHeapUptime; }
     size_t getMinimumHeap10min() { return _minimumHeap10min; }
     size_t getHeapLowErrUptime() { return _HeapLowErrUptime; }
